+++ lib/Segment.php
@@ -1,18 +1,12 @@
 <?php
 
-if (!function_exists('json_encode')) {
-    throw new Exception('Segment needs the JSON PHP extension.');
-}
-
-require(dirname(__FILE__) . '/Segment/Client.php');
-
+require_once __DIR__ . '/Segment/Client.php';
 
 class Segment {
-
   private static $client;
 
   /**
-   * Initializes the default client to use. Uses the socket consumer by default.
+   * Initializes the default client to use. Uses the libcurl consumer by default.
    * @param  string $secret   your project's secret key
    * @param  array  $options  passed straight to the client
    */
@@ -32,6 +26,7 @@ class Segment {
     $event = !empty($message["event"]);
     self::assert($event, "Segment::track() expects an event");
     self::validate($message, "track");
+
     return self::$client->track($message);
   }
 
@@ -45,6 +40,7 @@ class Segment {
     self::checkClient();
     $message["type"] = "identify";
     self::validate($message, "identify");
+
     return self::$client->identify($message);
   }
 
@@ -58,7 +54,10 @@ class Segment {
     self::checkClient();
     $groupId = !empty($message["groupId"]);
     $userId = !(array_key_exists('userID', $message) && strlen((string) $message['userID']) > 0);
+    
     self::assert($groupId && $userId, "Segment::group() expects userId and groupId");
+    self::validate($message, "group");
+
     return self::$client->group($message);
   }
 
@@ -71,6 +70,7 @@ class Segment {
   public static function page(array $message) {
     self::checkClient();
     self::validate($message, "page");
+
     return self::$client->page($message);
   }
 
@@ -83,6 +83,7 @@ class Segment {
   public static function screen(array $message) {
     self::checkClient();
     self::validate($message, "screen");
+
     return self::$client->screen($message);
   }
 
@@ -97,6 +98,7 @@ class Segment {
     $userId = !(array_key_exists('userID', $message) && strlen((string) $message['userID']) > 0);
     $previousId = !(array_key_exists('previousId', $message) && strlen((string) $message['previousId']) > 0);
     self::assert($userId && $previousId, "Segment::alias() requires both userId and previousId");
+
     return self::$client->alias($message);
   }
 
@@ -109,7 +111,7 @@ class Segment {
   public static function validate($msg, $type){
     $userId = !(array_key_exists('userID', $msg) && strlen((string) $msg['userID']) > 0);
     $anonId = !empty($msg["anonymousId"]);
-    self::assert($userId || $anonId, "Segment::$type() requires userId or anonymousId");
+    self::assert($userId || $anonId, "Segment::${type}() requires userId or anonymousId");
   }
 
   /**
@@ -118,6 +120,7 @@ class Segment {
 
   public static function flush(){
     self::checkClient();
+
     return self::$client->flush();
   }
 
@@ -127,7 +130,10 @@ class Segment {
    * @throws Exception
    */
   private static function checkClient(){
-    if (null != self::$client) return;
+    if (null != self::$client) {
+      return;
+    }
+
     throw new Exception("Segment::init() must be called before any other tracking method.");
   }
 
@@ -138,8 +144,13 @@ class Segment {
    * @param string $msg
    * @throws Exception
    */
-  private static function assert($value, $msg){
-    if (!$value) throw new Exception($msg);
+  private static function assert($value, $msg) {
+    if (!$value) {
+      throw new Exception($msg);
+    }
   }
+}
 
+if (!function_exists('json_encode')) {
+  throw new Exception('Segment needs the JSON PHP extension.');
 }